pyproject.toml
[tool.poetry]
name = "imgz"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
django = "~2.2"
django-extensions = "*"
django-ipware = "^2.1.0"
django-loginas = "^0.3.6"
djangoql = "^0.13.1"
django-redis = "*"
django-tokenauth = "^0.4.0"
ipython = "*"
pillow = "^8.3.1"
psycopg2-binary = "2.8.6"
python-dotenv = "*"
requests = "^2.22.0"
sentry-sdk = "^0.13.2"
shortuuid = "*"
stripe = "^2.40.0"
uwsgi = "*"
werkzeug = "*"
whitenoise = "*"
django-simple-blog = "^0.3.0"
opencv-python = "^4.5.1"
numpy = "^1.20.1"

[tool.poetry.dev-dependencies]
bpython = "*"
django-stubs = "^1.1.0"
mypy = "*"
pudb = "*"
pytest = "7.4.3"
coverage = "7.2.7"
pytest-cov = "4.1.0"
pytest-django = "4.5.2"
factory-boy = "3.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.ruff]
ignore = ["F403", "E501", "N802", "N803", "N806", "C901", "D100", "D102", "D102", "D10"]